[NSE-451] backport patches to 1.2 (#448)
* Minor: Remove debug code for PR#387 (#431)

* [NSE-436] Fix for Arrow Data Source test suite (#437)

Closes #436

* [NSE-254]Solve the redundant arrow library issue (#440)

* [NSE-254]Issue0410 jar size (#441)

* [NSE-254]Solve the redundant arrow library issue

* Remove mvn jar plugin 3.2.0

* fix packaging (#442)



* [NSE-207] Fix aggregate and refresh UT test script (#426)

* fix expr id difference in Partial Aggregate

* add fallback to window

* refresh ut

* use stol instead of stoi

* enable ut full test

* [NSE-429] TPC-DS Q14a/b get slowed down within setting spark.oap.sql.columnar.sortmergejoin.lazyread=true (#432)

* wip

* debug commit

* wip

* discard proxy pattern

* fix

* fix

* fix

* fix unexpected O(n^2) time

* fix1

* fix2

* Add a bunch of caches; Adjust prefetch batch length to 16

* format

* fix

* disable spark.oap.sql.columnar.sortmergejoin.lazyread by default

* fix

+++ native-sql-engine/cpp/src/codegen/arrow_compute/ext/hash_aggregate_kernel.cc
@@ -492,7 +492,7 @@ class HashAggregateKernel::Impl {
                  0) {
         auto res_type_list = {result_field_list[result_id]};
         result_id += 1;
-        int arg = std::stoi(action_name_list[action_id].substr(20));
+        int arg = std::stol(action_name_list[action_id].substr(20));
         RETURN_NOT_OK(MakeCountLiteralAction(ctx_, arg, res_type_list, &action));
       } else if (action_name_list[action_id].compare("action_stddev_samp_partial") ==
                  0) {
@@ -650,7 +650,7 @@ class HashAggregateKernel::Impl {
       } else if (action_name.compare(0, 20, "action_countLiteral_") == 0) {
         auto res_type_list = {result_field_list[result_id]};
         result_id += 1;
-        int arg = std::stoi(action_name.substr(20));
+        int arg = std::stol(action_name.substr(20));
         RETURN_NOT_OK(MakeCountLiteralAction(ctx_, arg, res_type_list, &action));
       } else if (action_name.compare("action_stddev_samp_partial") == 0) {
         auto res_type_list = {result_field_list[result_id],
